requirements.txt
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
nfl_data_py==0.3.2
pyarrow==17.0.0
scipy==1.13.1
